checkers/neftetochka/requirements.txt
checklib
grpcio
pycryptodome
protobuf
websocket-client==1.6.4
gpxpy
